Add implementation for US05 & US06 (#20)

+++ src/main/java/com/example/be_java_hisp_w25_g11/repository/buyer/IBuyerRepository.java
@@ -3,4 +3,4 @@ package com.example.be_java_hisp_w25_g11.repository.buyer;
 import com.example.be_java_hisp_w25_g11.entity.Buyer;
 import com.example.be_java_hisp_w25_g11.repository.ICrudRepository;
 
-public interface IBuyerRepository extends ICrudRepository <Buyer, Long> { }
+public interface IBuyerRepository extends ICrudRepository <Buyer, Integer> { }
