+++ src/main/java/com/example/be_java_hisp_w25_g11/repository/seller/ISellerRepository.java
@@ -6,6 +6,6 @@ import com.example.be_java_hisp_w25_g11.repository.ICrudRepository;
 import java.util.List;
 import java.util.Map;
 
-public interface ISellerRepository extends ICrudRepository <Seller, Long> {
+public interface ISellerRepository extends ICrudRepository <Seller, Integer> {
 
 }
